ISSUE #3626 - avoid rerendering all the comments on typing
