+++ frontend/src/v5/ui/routes/viewer/tickets/commentsPanel/commentsPanel.component.tsx
@@ -15,53 +15,27 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import CommentIcon from '@assets/icons/outlined/comment-outlined.svg';
-import SendIcon from '@assets/icons/outlined/send_message-outlined.svg';
-import FileIcon from '@assets/icons/outlined/file-outlined.svg';
 import { formatMessage } from '@/v5/services/intl';
-import { FormProvider, useForm } from 'react-hook-form';
 import { useParams } from 'react-router-dom';
-import { CurrentUserHooksSelectors, TicketsCardHooksSelectors, TicketsHooksSelectors } from '@/v5/services/selectorsHooks';
+import { TicketsCardHooksSelectors, TicketsHooksSelectors } from '@/v5/services/selectorsHooks';
 import { TicketsActionsDispatchers } from '@/v5/services/actionsDispatchers';
 import { modelIsFederation } from '@/v5/store/tickets/tickets.helpers';
 import { ScrollArea } from '@controls/scrollArea';
 import { FormattedMessage } from 'react-intl';
 import { IComment } from '@/v5/store/tickets/tickets.types';
-import DeleteIcon from '@assets/icons/outlined/close-outlined.svg';
 import { useEffect, useState } from 'react';
 import { ViewerParams } from '../../../routes.constants';
-import {
-	Accordion,
-	Comments,
-	Controls,
-	CharsCounter,
-	FileIconButton,
-	SendButton,
-	BottomSection,
-	MessageInput,
-	EmptyCommentsBox,
-	DeleteButton,
-	CommentReplyContainer,
-} from './commentsPanel.styles';
+import { Accordion, Comments, EmptyCommentsBox } from './commentsPanel.styles';
 import { Comment } from './comment/comment.component';
-import { MAX_MESSAGE_LENGTH, addReply, createMetadata, sanitiseMessage } from './comment/comment.helpers';
-import { CommentReply } from './comment/commentReply/commentReply.component';
+import { CreateCommentBox } from './createCommentBox/createCommentBox.component';
 
 export const CommentsPanel = ({ scrollPanelIntoView }) => {
 	const [commentReply, setCommentReply] = useState<IComment>(null);
-	const [isSubmittingMessage, setIsSubmittingMessage] = useState(false);
-	const formData = useForm<{ message: string }>({ mode: 'all' });
-	const messageInput = formData.watch('message');
-
 	const { teamspace, project, containerOrFederation } = useParams<ViewerParams>();
 	const isFederation = modelIsFederation(containerOrFederation);
 	const ticketId = TicketsCardHooksSelectors.selectSelectedTicketId();
 	const { comments = [], number } = TicketsHooksSelectors.selectTicketById(containerOrFederation, ticketId);
 
-	const currentUser = CurrentUserHooksSelectors.selectCurrentUser();
-
-	const commentReplyLength = commentReply ? addReply(createMetadata(commentReply), '').length : 0;
-	const charsCount = (messageInput?.length || 0) + commentReplyLength;
-	const charsLimitIsReached = charsCount >= MAX_MESSAGE_LENGTH;
 	const commentsListIsEmpty = comments?.length > 0;
 
 	const handleDeleteComment = (commentId) => {
@@ -98,34 +72,6 @@ export const CommentsPanel = ({ scrollPanelIntoView }) => {
 		);
 	};
 
-	const resetCommentBox = () => {
-		setCommentReply(null);
-		formData.reset();
-		setIsSubmittingMessage(false);
-	};
-
-	const createComment = async () => {
-		setIsSubmittingMessage(true);
-		let message = sanitiseMessage(messageInput);
-		if (commentReply) {
-			message = addReply(createMetadata(commentReply), message);
-		}
-		const newComment = {
-			author: currentUser.username,
-			message,
-		};
-		TicketsActionsDispatchers.createTicketComment(
-			teamspace,
-			project,
-			containerOrFederation,
-			ticketId,
-			isFederation,
-			newComment,
-			resetCommentBox,
-			() => setIsSubmittingMessage(false),
-		);
-	};
-
 	useEffect(() => {
 		if (!number) return;
 		TicketsActionsDispatchers.fetchTicketComments(
@@ -135,7 +81,6 @@ export const CommentsPanel = ({ scrollPanelIntoView }) => {
 			ticketId,
 			isFederation,
 		);
-		resetCommentBox();
 	}, [number]);
 
 	return (
@@ -144,61 +89,30 @@ export const CommentsPanel = ({ scrollPanelIntoView }) => {
 			Icon={CommentIcon}
 			onChange={scrollPanelIntoView}
 		>
-			<FormProvider {...formData}>
-				<ScrollArea autoHeight autoHeightMin={400} autoHeightMax={400}>
-					{commentsListIsEmpty && (
-						<Comments>
-							{comments.map((comment) => (
-								<Comment
-									{...comment}
-									key={comment._id}
-									onDelete={handleDeleteComment}
-									onReply={handleReplyToComment}
-									onEdit={handleEditComment}
-								/>
-							))}
-						</Comments>
-					)}
-					{!commentsListIsEmpty && (
-						<EmptyCommentsBox>
-							<FormattedMessage id="ticket.comments.empty" defaultMessage="No comments" />
-						</EmptyCommentsBox>
-					)}
-				</ScrollArea>
-				<BottomSection>
-					{commentReply && (
-						<CommentReplyContainer>
-							<CommentReply {...createMetadata(commentReply)} shortMessage />
-							<DeleteButton onClick={() => setCommentReply(null)}>
-								<DeleteIcon />
-							</DeleteButton>
-						</CommentReplyContainer>
-					)}
-					<MessageInput
-						name="message"
-						placeholder={formatMessage({
-							id: 'customTicket.panel.comments.leaveAComment',
-							defaultMessage: 'Leave a comment',
-						})}
-						inputProps={{
-							maxLength: Math.max(MAX_MESSAGE_LENGTH - commentReplyLength, 0),
-						}}
-					/>
-					<Controls>
-						<FileIconButton>
-							<FileIcon />
-						</FileIconButton>
-						<CharsCounter $error={charsLimitIsReached}>{charsCount}/{MAX_MESSAGE_LENGTH}</CharsCounter>
-						<SendButton
-							disabled={!messageInput?.trim()?.length || charsLimitIsReached}
-							onClick={createComment}
-							isPending={isSubmittingMessage}
-						>
-							<SendIcon />
-						</SendButton>
-					</Controls>
-				</BottomSection>
-			</FormProvider>
+			<ScrollArea autoHeight autoHeightMin={400} autoHeightMax={400}>
+				{commentsListIsEmpty && (
+					<Comments>
+						{comments.map((comment) => (
+							<Comment
+								{...comment}
+								key={comment._id}
+								onDelete={handleDeleteComment}
+								onReply={handleReplyToComment}
+								onEdit={handleEditComment}
+							/>
+						))}
+					</Comments>
+				)}
+				{!commentsListIsEmpty && (
+					<EmptyCommentsBox>
+						<FormattedMessage id="ticket.comments.empty" defaultMessage="No comments" />
+					</EmptyCommentsBox>
+				)}
+			</ScrollArea>
+			<CreateCommentBox
+				commentReply={commentReply}
+				setCommentReply={setCommentReply}
+			/>
 		</Accordion>
 	);
 };
