+++ frontend/src/v5/ui/routes/viewer/tickets/commentsPanel/commentsPanel.styles.ts
@@ -17,11 +17,7 @@
 
 import styled from 'styled-components';
 import { Accordion as AccordionBase } from '@controls/accordion/accordion.component';
-import { Typography } from '@controls/typography';
-import { FormTextAreaFixedSize } from '@controls/inputs/formInputs.component';
-import { Container as TextAreaContainer } from '@controls/inputs/textArea/textAreaFixedSize.styles';
 import { EmptyListMessage } from '@controls/dashedContainer/emptyListMessage/emptyListMessage.styles';
-import { SubmitButton } from '@controls/submitButton';
 
 export const Accordion = styled(AccordionBase)`
 	&& {
@@ -40,91 +36,6 @@ export const Comments = styled.div`
 	flex-direction: column;
 	padding: 0 13px 10px 10px;
 	overflow-x: hidden;
-`;
-
-export const BottomSection = styled.section`
-	display: flex;
-	flex-direction: column;
-	border: solid 0 ${({ theme }) => theme.palette.secondary.lightest};
-	border-top-width: 1px;
-	padding: 0 15px 11px;
-`;
-
-export const CommentReplyContainer = styled.div`
-	position: relative;
-	margin-top: 11px;
-`;
-
-export const DeleteButton = styled.div`
-	position: absolute;
-	top: -7px;
-	right: -7px;
-	height: 24px;
-	width: 24px;
-
-	display: flex;
-	align-items: center;
-	justify-content: center;
-
-	background: ${({ theme }) => theme.palette.primary.contrast};
-	border-radius: 100%;
-	box-shadow: 0 3px 5px rgba(0, 0, 0, 0.15);
-	cursor: pointer;
-
-	svg {
-		width: 10px;
-	}
-`;
-
-export const MessageInput = styled(FormTextAreaFixedSize)`
-	.MuiInputBase-multiline {
-		padding: 8px 0 6px;
-		line-height: 16px;
-	}
-
-	${TextAreaContainer} {
-		border: none;
-		box-shadow: none;
-		padding: 0;
-	}
-`;
-
-export const Controls = styled.div`
-	width: 100%;
-	display: flex;
-	align-items: center;
-	justify-content: flex-start;
-	padding-top: 6px;
-	box-sizing: border-box;
-	box-shadow: 0 0 9px 7px ${({ theme }) => theme.palette.primary.contrast};
-	position: relative;
-`;
-
-export const CharsCounter = styled(Typography).attrs({
-	variant: 'body1',
-})<{ $error?: boolean }>`
-	margin: 1px 0 0 11px;
-	font-weight: 500;
-	color: ${({ theme: { palette }, $error }) => ($error ? palette.error.main : palette.base.lighter)};
-`;
-
-export const FileIconButton = styled.div`
-	cursor: pointer;
-	display: flex;
-	padding: 4px;
-	color: ${({ theme }) => theme.palette.secondary.main};
-`;
-
-export const SendButton = styled(SubmitButton).attrs({
-	color: 'primary',
-	variant: 'contained',
-})`
-	margin: 0 0 0 auto;
-	border-radius: 50%;
-	padding: 0;
-	min-width: unset;
-	width: 34px;
-	height: 34px;
 `;
 
 export const EmptyCommentsBox = styled(EmptyListMessage)`
